% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_tree.R
\name{predict_motr_bart}
\alias{predict_motr_bart}
\title{Title}
\usage{
predict_motr_bart(object, newdata, type = c("all", "median", "mean"))
}
\arguments{
\item{object}{The fit object returned by the function \code{motr_bart}.}

\item{newdata}{The design matrix that will be used to predict the continuous outcome for.}

\item{type}{Should the predicted values be the \code{"mean"} or \code{"median"} of the posterior draws? It is also possible to set \code{"all"}, where no function is applied and the posteriors draws are returned for each MCMC iteration.}
}
\value{
If \code{type = "mean"} or \code{type = "median"}, then a vector with predicted values is returned. If \code{type = "all"}, an n times npost is returned, where n is the number of row of \code{newdata}.
}
\description{
Title
}
